pyproject.toml
[tool.poetry]
package-mode = false
name = "codebase"
version = "1.0.1"
description = "Project"
license = "UNLICENSED"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.34.2"
fastapi-limiter = "^0.1.6"
fastapi = {extras = ["all"], version = "^0.115.12"}
websockets = "^15.0.1"
python-socketio = "^5.11.0"
python-multipart = "^0.0.20"
python-ulid = "^2.3.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.270.0"}
psutil = "^5.9.8"
click = "^8.1.7"
toml = "^0.10.2"
jinja2 = "^3.1.6"
apscheduler = "^3.10.4"
aiosmtplib = "^3.0.2"
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"
redis = "^6.0.0"
rq = "^1.16.1"
psycopg2 = "^2.9.10"
sqlmodel = "^0.0.24"
alembic = "^1.13.2"
motor = "^3.3.2"
faker = "^33.1.0"
scrapy = "^2.13.0"
pandas = "^2.2.3"
numpy = "^2.2.6"
scikit-learn = "^1.6.1"
keras = "^3.10.0"
transformers = "^4.52.1"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.9"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
